backend/requirements.txt
django
psycopg2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow